Use Smoothie Charts for graphs
The goal here is to keep the JS code agnostic regarding the HTML
page being processed.
In the template, a "canvas" html node can be added, with the following
attributes:
* use the "smoothiechart" class
* have a "data-graph-variable" custom attribute, whose value is the id
  of the html node where the value can be retrieved.

The JS code will detect all the "smoothiechart" class usage, and create
a SmoothieChart for each canvas, linking the TimeSerie of the chart to
the value pointed by "data-graph-variable".

Note: the variable is supposed to be between 0 and 100, like a percentage.
Note2: a hidden html node can be used to store the value. The Memory
       canvas uses this trick to plot the memory percentage.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>